feat: add ui store and integrate
